requirements.txt
fastapi
uvicorn
kafka-python
pymongo
pydantic
python-dotenv